+++ docs/src/upgrade/release-4.x.x.asciidoc
@@ -491,57 +491,15 @@ unwrap('hello');    // 'hello'
 
 ==== Subgraph Support in GLVs
 
-The GraphBinary 4.0 `Graph` type (`0x10`) is now handled by `gremlin-javascript`, `gremlin-dotnet`, and `gremlin-go`,
-so a `subgraph()`-step result returned to the client is deserialized into a `Graph` data container populated with the
-captured vertices and edges. Reference documentation: <<subgraph-step>>.
+All GLVs now support the `subgraph()` step. Previously, calling `subgraph()` from a GLV produced an unknown-type error
+because the variant could not interpret the `Graph` payload that the server returned. Applications can now extract a
+portion of a remote graph as part of a normal traversal and inspect its vertices and edges directly from the client,
+without having to re-issue queries to reconstruct the result. See: <<subgraph-step>>.
 
-The returned `Graph` is a detached snapshot of the captured vertices and edges. Mutating its collections has no
-effect on the source graph.
-
-* `gremlin-javascript`: `graph.vertices: Map<any, Vertex>` and `graph.edges: Map<any, Edge>`
-* `gremlin-dotnet`: `graph.Vertices: IDictionary<object, Vertex>` and `graph.Edges: IDictionary<object, Edge>`
-* `gremlin-go`: `graph.Vertices: map[interface{}]*Vertex` and `graph.Edges: map[interface{}]*Edge`
-
-IMPORTANT: The deserialized `Graph` is a detached data container, not a traversable `Graph` instance. It cannot be
-passed to `traversal().with(...)` from a GLV. To re-query elements from the original graph, extract their ids and
-call `g.V(id)` or `g.E(id)` against the original `GraphTraversalSource`.
-
-Example (`gremlin-javascript`):
-
-[source,javascript]
-----
-const sg = await g.E().hasLabel('knows').subgraph('sg').cap('sg').next();
-for (const [id, vertex] of sg.vertices) {
-  console.log(id, vertex.label);
-}
-for (const [id, edge] of sg.edges) {
-  console.log(`${edge.outV.id} -${edge.label}-> ${edge.inV.id}`);
-}
-----
-
-Example (`gremlin-dotnet`):
-
-[source,csharp]
-----
-var sg = (Graph) await g.E().HasLabel("knows").Subgraph("sg").Cap<object>("sg").Next();
-foreach (var entry in sg.Vertices)
-    Console.WriteLine($"{entry.Key} {entry.Value.Label}");
-----
-
-Example (`gremlin-go`):
-
-[source,go]
-----
-result, _ := g.E().HasLabel("knows").Subgraph("sg").Cap("sg").Next()
-sg := result.Data.(*gremlingo.Graph)
-for id, v := range sg.Vertices {
-    fmt.Println(id, v.Label)
-}
-----
-
-The `Graph` string representation in `gremlin-python`, `gremlin-javascript`, `gremlin-dotnet`, and `gremlin-go` now
-renders as `graph[vertices:N edges:M]`, matching the format used by Java's `AbstractTinkerGraph.toString()` for easier
-debugging.
+In the GLVs, the result is a detached snapshot of the captured vertices and edges, not a traversable `Graph` instance.
+It cannot be passed to `traversal().with(...)`, and mutating its collections has no effect on the source graph. To
+re-query elements against the original graph, extract their ids and call `g.V(id)` or `g.E(id)` against the original
+`GraphTraversalSource`.
 
 === Upgrading for Providers
